ESDEV-4156 Convert all errors to exception
Typically PHP errors do not use return code as an indication of
error, meanwhile exceptions use return code to notify end user
that the exception was indeed thrown. Converting errors to
exceptions helps to catch these errors in a form of return code
which can then be used e.g. for setup process to identify if view
regeneration went successfully.